Fix processing /proc/zoneinfo. Split node and zone stats.
1. File /proc/zoneinfo contains information about two different entities:
node stats and zone stats. Kernel developers decided not to create
a dedicated file in /proc for node stats but mix-in those stats
in zoneinfo (linux-kernel commit e2ecc8a79ed49f7838b4fdf352c4c48cec9424ac)
under the first populated zone in each node.

Current implementation /proc/zoneinfo parsing is incorrect as it tries
to merge attributes of different entities into a single frankenstein
structure Zoneinfo, which is a bad design pattern. Moreover, due to
this decision, there is a bug in parsing data of the first zone in node.

Given the /proc/zoneinfo lines:

Node 0, zone      DMA
  per-node stats
      ...
  pages free x
      ...

the resulting Zoneinfo element will have 'Zone' field empty (that is zone
name "DMA" is discarded) as the current code interprets "per-node stats"
line as a marker that all of the following lines in this section will be
related to node stats only. Which is incorrect, as "DMA" zone stats
("pages free" and below) will follow next.

2. Renamed/removed some stats:
   "nr_mlock_stack" - cannot find any reference to it in kernel git
       nor can I see it in current kernels' /proc/zoneinfo (and test
       fixtures)
   "nr_mlock" - added instead

   added some stats that appeared in current kernels.

Signed-off-by: Sergei Turchanov <turchanov@farpost.com>